bot/requirements.txt
aiogram==3.17.0
python-dotenv==1.0.1
requests==2.32.3
aiohttp==3.11.11